Fix bug in ComputeGradientFeatures

+++ histomicstk/features/ComputeGradientFeatures.py
@@ -126,7 +126,7 @@ def ComputeGradientFeatures(im_label, im_intensity,
         fdata.at[i, 'Gradient.Mag.HistEnergy'] = np.sum(prob**2)
 
         bw_canny = cannyG[rprops[i].coords[:, 0], rprops[i].coords[:, 1]]
-        canny_sum = np.sum(bw_canny)
+        canny_sum = np.sum(bw_canny).astype('float')
         fdata.at[i, 'Gradient.Canny.Sum'] = canny_sum
 
         fdata.at[i, 'Gradient.Canny.Mean'] = canny_sum / len(pixelGradients)
